Fixed averageSize int overflow.
In the previous variant max average size could be only Integer.MAX_VALUE / 16 = 134217727 ~ 64 Mb String (java 8), because of row:
averageSize = (size + 15 * averageSize) / 16;

See the issue: https://github.com/h2database/h2database/issues/2967

+++ h2/src/main/org/h2/mvstore/type/ObjectDataType.java
@@ -1528,7 +1528,7 @@ public class ObjectDataType extends BasicDataType<Object> {
      */
     static class SerializedObjectType extends AutoDetectDataType<Object> {
 
-        private int averageSize = 10_000;
+        private long averageSize = 10_000;
 
         SerializedObjectType(ObjectDataType base) {
             super(base, TYPE_SERIALIZED_OBJECT);
@@ -1574,7 +1574,7 @@ public class ObjectDataType extends BasicDataType<Object> {
         public int getMemory(Object obj) {
             DataType<Object> t = getType(obj);
             if (t == this) {
-                return averageSize;
+                return (int) averageSize;
             }
             return t.getMemory(obj);
         }
